Completed Processor.java and GetData2.getData()
Processor processes data from GetData1.getData() to
construct a map and then pass it to GetData2.getData().

GetData2.getData() fetch data from table2 based on the
key of the passed map. Finally it will determine which
phone number is going to be added to the list for notification.

TODO: improve README and TEST the program, and parameter modificatons.